FUNCZ-349 Synapse property names end in _mu if they signify a mean value
This should clear up some naming confusion. E.g., `conductance` is
really the value for the mean of the distribution the real `conductance`
value is drawn from.

I think this change would make it more clear to the user that these
values are means, and lead to less confusion w.r.t. the final output
value.